# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Veil Panel API
  version: "0.6.3"
  description: |
    HTTP management API exposed by the Veil Panel (`veil serve`).

    ## Authentication model
    Veil is fail-closed for public Panel exposure. A non-loopback `--listen`
    address requires native TLS, API token authentication (`--auth-token` or
    `VEIL_API_TOKEN`), user/session authentication, and authenticated metrics.
    Caddy exposure requires user/session authentication and authenticated
    metrics even though the Veil listener remains on loopback.

    An unconfigured local loopback listener exposes only the first-run setup
    flow needed to create the initial administrator. Setup is unavailable
    through direct or Caddy exposure.

    `/api/*` routes accept either `X-Veil-Token`, `Authorization: Bearer`, or
    a `veil_session` cookie created by `/api/auth/login`. Mutating requests
    made with a cookie session must also send `X-CSRF-Token`, and only `admin`
    sessions may mutate management state. The sole viewer mutation exception
    is `POST /api/auth/locale`, which updates only the authenticated user's own
    language preference and rejects static API-token callers.

    Browser sessions survive Panel restarts. Only hashes of the session cookie
    and CSRF token are persisted. Sessions expire after 30 minutes idle or 24
    hours absolute and are revoked when a user's password, role, or existence
    changes.

...
    until a valid session exists.

...
  license:
...
servers:
...
security:
  - veilToken: []
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /api/auth/sessions:
    get:
      x-veil-role: admin-metadata
      tags: [auth]
      summary: List active browser sessions
      description: Requires an admin token or admin session.
      security:
        - veilToken: []
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Active sessions with stable revocation IDs. The current session is sorted first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionInfo"
              examples:
                sessions:
                  value:
                    - id: 0123456789abcdef
                      username: admin
                      role: admin
                      createdAt: "2026-06-05T12:00:00Z"
                      lastSeenAt: "2026-06-05T12:10:00Z"
                      idleExpiresAt: "2026-06-05T12:40:00Z"
                      expiresAt: "2026-06-06T12:00:00Z"
                      userAgent: Mozilla/5.0
                      remoteAddr: 192.0.2.10
                      current: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      x-veil-role: admin-mutation
      tags: [auth]
      summary: Revoke an active browser session
      description: Requires an admin token or admin session. Cookie sessions must include `X-CSRF-Token`.
      security:
        - veilToken: []
        - bearerAuth: []
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionDeleteRequest"
            examples:
              revoke:
                value:
                  id: 0123456789abcdef
      responses:
        "200":
          description: Session was revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              examples:
                ok:
                  value:
                    success: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    veilToken:
      type: apiKey
      in: header
      name: X-Veil-Token
    bearerAuth:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: veil_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
  responses:
    BadRequest:
      description: Request body, query string, or path failed validation.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorText"
          examples:
            invalid:
              value: "invalid JSON\n"
    Unauthorized:
      description: Missing or invalid token/session.
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Present on authenticated API challenges.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorText"
          examples:
            missing:
              value: "missing or invalid API token or session\n"
    Forbidden:
      description: Authenticated principal is not allowed to perform the operation.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorText"
          examples:
            csrf:
              value: "invalid or missing CSRF token\n"
    NotFound:
      description: Resource not found.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorText"
          examples:
            not-found:
              value: "404 page not found\n"
...
  schemas:
...
    ErrorText:
      type: string
      description: Plain text error emitted by `writeError`; usually a single message line ending in `\n`.
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    SessionInfo:
      type: object
      required: [id, username, role, createdAt, lastSeenAt, idleExpiresAt, expiresAt, current]
      properties:
        id:
          type: string
          description: Stable short hash used for session revocation; the raw bearer cookie is never returned.
          minLength: 16
          maxLength: 16
        username:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        createdAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
        idleExpiresAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        userAgent:
          type: string
        remoteAddr:
          type: string
        current:
          type: boolean
          description: True when this entry matches the caller's `veil_session` cookie.
    SessionDeleteRequest:
      type: object
      required: [id]
      additionalProperties: false
      properties:
        id:
          type: string
          minLength: 16
          maxLength: 16
    UserRole:
      type: string
      enum: [admin, viewer]
...